+++ contracts/mock/NoxComputeV2Mock.sol
@@ -9,7 +9,7 @@ import {NoxCompute} from "../NoxCompute.sol";
  * Adds a version() function to verify the upgrade was applied.
  */
 contract NoxComputeV2Mock is NoxCompute {
-    constructor(uint24 cuPerOperation) NoxCompute(cuPerOperation) {}
+    constructor(uint8 cuPerOperation) NoxCompute(cuPerOperation) {}
 
     /**
      * Returns the contract version.
